[shape_poly] Fix handling of shape polymorphism for pallas_call_batching
The previous code was not handling the case of a symbolic batch
dimension (one of the most common uses of shape polymorphism)

I also moved the symbolic shape tests from pallas_test to export_pallas_test,
along with all the other symbolic shape tests for Pallas.
