+++ tests/pallas/pallas_test.py
@@ -2608,120 +2608,6 @@ class PallasCallNamedGridTest(ptu.PallasTest):
     )
 
 
-class SymbolicPallasTest(ptu.PallasTest):
-
-  def test_simple_symbolic_matmul_export(self):
-    if jtu.test_device_matches(["gpu"]):
-      self.skipTest("Not supported on GPU.")
-
-    def sym_matmul(x, y, symbolic_grid):
-      symbolic_grid = symbolic_grid.shape[0]
-      symbolic_x_0 = x.shape[0] // symbolic_grid
-      symbolic_y_1 = y.shape[1] // symbolic_grid
-
-      def x_ref_block_spec_mapping(i, j):
-        return (i, 0)
-
-      def y_ref_block_spec_mapping(i, j):
-        return (0, j)
-
-      def sym_matmul_kernel(x_ref, y_ref, z_ref):
-        z_ref[...] = x_ref[...] @ y_ref[...]
-
-      return pl.pallas_call(
-          sym_matmul_kernel,
-          out_shape=jax.ShapeDtypeStruct((symbolic_x_0, symbolic_y_1), x.dtype),
-          grid_spec=pltpu.PrefetchScalarGridSpec(
-              num_scalar_prefetch=0,
-              in_specs=[
-                  pl.BlockSpec(
-                      (symbolic_x_0, x.shape[1]), x_ref_block_spec_mapping
-                  ),
-                  pl.BlockSpec(
-                      (y.shape[0], symbolic_y_1),
-                      y_ref_block_spec_mapping,
-                  ),
-              ],
-              out_specs=pl.BlockSpec(
-                  (symbolic_x_0, symbolic_y_1),
-                  lambda i, j: (i, j),
-              ),
-              grid=(symbolic_grid, symbolic_grid),
-          ),
-      )(x, y)
-
-    a, b, c, d, e = jax.export.symbolic_shape(
-        "m_dim, k_dim, n_dim, grid_size, unused_dim",
-        constraints=(
-            "mod(floordiv(m_dim, grid_size), 8) == 0",
-            "mod(k_dim, 128) == 0",
-            "mod(floordiv(n_dim, grid_size), 128) == 0",
-        ),
-    )
-    x = jax.ShapeDtypeStruct((a, b), jax.numpy.float32)
-    y = jax.ShapeDtypeStruct((b, c), jax.numpy.float32)
-
-    dummy_d = jax.ShapeDtypeStruct((d, e), jax.numpy.float32)
-
-    exported_module = pl.lower_as_mlir(
-        jax.jit(sym_matmul), x, y, dummy_d, dynamic_shapes=True
-    )
-    assert exported_module is not None
-    self.assertIn(
-        "%arg0: tensor<?x?xf32> loc(unknown), %arg1: tensor<?x?xf32>"
-        " loc(unknown), %arg2: tensor<?x?xf32>",
-        str(exported_module),
-    )
-    x = jax.ShapeDtypeStruct((128, 1024), jax.numpy.float32)
-    y = jax.ShapeDtypeStruct((1024, 512), jax.numpy.float32)
-    dummy_d = jax.ShapeDtypeStruct((1, 1), jax.numpy.float32)
-    exported_module = pl.lower_as_mlir(
-        jax.jit(sym_matmul), x, y, dummy_d, dynamic_shapes=False
-    )
-    assert exported_module is not None
-    self.assertIn(
-        "call @sym_matmul(%arg0, %arg1)",
-        str(exported_module),
-    )
-
-  def test_pallas_shape_poly_no_cache_collision(self):
-
-    def kernel(x, y):
-      y[:] = x[:]
-
-    f = self.pallas_call(
-        kernel,
-        out_shape=jax.ShapeDtypeStruct((8, 128), jnp.float32),
-    )
-    f = jax.vmap(f)
-
-    x1_shape = jax.ShapeDtypeStruct(
-        jax.export.symbolic_shape('b1, 8, 128'), jnp.float32
-    )
-    exported_module1 = pl.lower_as_mlir(jax.jit(f), x1_shape, dynamic_shapes=True)
-    self.assertIn("(b1, 8, 128)", str(exported_module1))
-    x2_shape = jax.ShapeDtypeStruct(
-        jax.export.symbolic_shape('b2, 8, 128'), jnp.float32
-    )
-    exported_module2 = pl.lower_as_mlir(jax.jit(f), x2_shape, dynamic_shapes=True)
-    self.assertIn("(b2, 8, 128)", str(exported_module2))
-
-  def test_cdiv(self):
-    def kernel(x, y):
-      c = pl.cdiv(x.shape[0], 2)
-      assert c == (x.shape[0] + 1) // 2
-      assert c == y.shape[0]
-      y[:] = x[:c]
-
-    m, = jax.export.symbolic_shape('m')
-    x_shape = jax.ShapeDtypeStruct((m, 128), jnp.float32)
-    y_shape = jax.ShapeDtypeStruct(((m + 1) // 2, 128), jnp.float32)
-    f = self.pallas_call(kernel, out_shape=y_shape)
-
-    exported_module = pl.lower_as_mlir(jax.jit(f), x_shape, dynamic_shapes=True)
-    self.assertIn("(m, 128)", str(exported_module))
-
-
 class PallasCallNamedGridInterpretTest(PallasCallNamedGridTest):
   INTERPRET = True
